feat(main-api): seperate tests code
